Use ExternalPackage and CPM together if CMocka not installed
If cmocka was not installed there was a CMake error
due to the packaging target created by CMocka when being built. This
technique first downloads the package with CPM but does not build it.
Then ExternalPackage_Add installs CMocka in the build directory.

This way, CPM is able to cache of the CMocka source so it doesn't have
to fetch it from a server after every distclean.